requirements.txt
ipykernel==6.29.3
openai==1.59.5
python-dotenv==1.0.1
yfinance==0.2.54
pandas==2.2.3
matplotlib==3.9.4
colorama==0.4.6